Add smoke test
Adds basic smoke test that creates a single workspace containing a
single deployment, service and ingress, and verifies the ingress
reconciles as expected.

Refactors the performance test to use the same basic ingress test for
both smoke and performance test runs.
