+++ test/support/ingress.go
@@ -1,5 +1,3 @@
-//go:build e2e || performance
-
 /*
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
